backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
requests==2.32.3
python-dotenv==1.0.1
prometheus-client==0.17.0
pytest==8.4.0
httpx==0.28.1
redis==5.2.1

# database libraries
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9
